added interior nodes
